Delegates as a separate method
